Add update_only in SubscriberStateTable.
If the flag is set, the API will return empty attribute (instead of full
list of attributes) when there is no change for SET requests. This will
be useful when CONFIG DB is updated with no actual change. After setting
the flag, the user of the SubscriberStateTable will do no-op when empty attribute is return.